IntegerColumn: removed unneeded cast

+++ Library/src/uk/ac/ucl/excites/sapelli/storage/model/columns/IntegerColumn.java
@@ -298,7 +298,7 @@ public class IntegerColumn extends NumberColumn<Long>
 	 */
 	public void storeValue(ValueSet<?> valueSet, Integer value) throws IllegalArgumentException, NullPointerException
 	{
-		storeValue(valueSet, (Long) convert(value));
+		storeValue(valueSet, convert(value));
 	}
 
 	/**
